# openApi/server/openapi_server/openapi/openapi.yaml
openapi: 3.0.2
info:
  description: An OpenAPI specification for Hivemind
  title: Hivemind OpenAPI Specification
  version: 0.0.1
servers:
- description: Hivemind local API Service
  url: http://localhost:8080/
paths:
  /#bridge.account_notifications:
    post:
      description: Lists notifications for given account
      operationId: bridge_account_notifications
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountNotificationsRequest'
        description: 'required account, optional: min_score, last_id, limit'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/AccountNotifications'
                - $ref: '#/components/schemas/ErrorMessage'
          description: list of notifications
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.does_user_follow_any_lists:
    post:
      description: Tells if given observer follows any blacklist or mute list
      operationId: bridge_does_user_follow_any_lists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoesUserFollowAnyListsRequest'
        description: required observer
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/DoesUserFollowAnyLists'
                - $ref: '#/components/schemas/ErrorMessage'
          description: Answer whether the observer follow any list
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.get_account_posts:
    post:
      description: Lists posts related to given account in selected way.
      operationId: bridge_get_account_posts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetAccountPostsRequest'
        description: 'required: sort, account, optional: start_author, start_permlink,
          limit, observer'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/GetAccountPosts'
                - $ref: '#/components/schemas/ErrorMessage'
          description: |
            List of posts
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.get_community:
    post:
      description: Gets community
      operationId: bridge_get_community
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityRequest'
        description: community name and optional observer
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/Community'
                - $ref: '#/components/schemas/ErrorMessage'
          description: Community result
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.get_community_context:
    post:
      description: Gets community context
      operationId: bridge_get_community_context
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityContextRequest'
        description: community name and account for context
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/CommunityContext'
                - $ref: '#/components/schemas/ErrorMessage'
          description: Community context result
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.get_discussion:
    post:
      description: Gives flattened discussion tree starting at given post.
      operationId: bridge_get_discussion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetDiscussionRequest'
        description: 'required: author, permlink, optional: observer'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/GetDiscussion'
                - $ref: '#/components/schemas/ErrorMessage'
          description: |
            List of discussion post
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.get_follow_list:
    post:
      description: For given observer gives directly blacklisted/muted accounts or
        list of blacklists/mute lists followed by observer
      operationId: bridge_get_follow_list
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetFollowListRequest'
        description: 'required: observer, optional: follow_type'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/GetFollowList'
                - $ref: '#/components/schemas/ErrorMessage'
          description: List of blacklisted/ muted lists
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.get_payout_stats:
    post:
      description: Lists communities ordered by payout with stats (total payout, number
        of posts and authors).
      operationId: bridge_get_payout_stats
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetPayoutStatsRequest'
        description: 'optional: limit'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/GetPayoutStats'
                - $ref: '#/components/schemas/ErrorMessage'
          description: List of communities with stats.
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.get_post:
    post:
      description: Gives single selected post.
      operationId: bridge_get_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetPostRequest'
        description: 'required: author, permlink, optional: observer'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/GetPost'
                - $ref: '#/components/schemas/ErrorMessage'
          description: Selected post
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.get_post_header:
    post:
      description: Gives very basic information on given post.
      operationId: bridge_get_post_header
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetPostHeaderRequest'
        description: 'required: author, permlink'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/GetPostHeader'
                - $ref: '#/components/schemas/ErrorMessage'
          description: Selected post description
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.get_profile:
    post:
      description: Gets profile
      operationId: bridge_get_profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetProfileRequest'
        description: 'required account, optional: observer'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/Profile'
                - $ref: '#/components/schemas/ErrorMessage'
          description: profile information
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.get_ranked_posts:
    post:
      description: Lists posts depending on given ranking criteria and filters.
      operationId: bridge_get_ranked_posts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetRankedPostsRequest'
        description: 'required: sort, optional: start_author, start_permlink, limit,
          tag, observer'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/GetRankedPosts'
                - $ref: '#/components/schemas/ErrorMessage'
          description: |
            Selected post
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.get_relationship_between_accounts:
    post:
      description: Tells what relations connect given accounts from the perspective
        of first account.
      operationId: bridge_get_relationship_between_accounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetRelationshipBetweenAccountsRequest'
        description: 'required: account1, account2, optional: observer'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/GetRelationshipBetweenAccounts'
                - $ref: '#/components/schemas/ErrorMessage'
          description: Account relations
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.list_all_subscriptions:
    post:
      description: Lists all community contexts from communities given account is
        subscribed to.
      operationId: bridge_list_all_subscriptions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListAllSubscriptionsRequest'
        description: points to valid account (not necessarily subscibed to any community)
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ListAllSubscriptions'
                - $ref: '#/components/schemas/ErrorMessage'
          description: all subscriptions list
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.list_communities:
    post:
      description: Gets community
      operationId: bridge_list_communities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListCommunitesRequest'
        description: 'optional parameters: last, limit, query, sort, observer'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ListCommunity'
                - $ref: '#/components/schemas/ErrorMessage'
          description: list of Communities
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.list_community_roles:
    post:
      description: Gets community
      operationId: bridge_list_community_roles
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListCommunityRolesRequest'
        description: community name and optional observer
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ListCommunityRoles'
                - $ref: '#/components/schemas/ErrorMessage'
          description: list of Community roles
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.list_pop_communities:
    post:
      description: Gets community
      operationId: bridge_list_pop_communities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListPopCommunitesRequest'
        description: 'optional parameter: limit'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ListPopCommunities'
                - $ref: '#/components/schemas/ErrorMessage'
          description: list of Communities
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /#bridge.list_subscribers:
    post:
      description: list subscribers
      operationId: bridge_list_subscribers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListSubscribersRequest'
        description: 'required community, optional: last, limit'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ListSubscribers'
                - $ref: '#/components/schemas/ErrorMessage'
          description: list of Communities
      x-openapi-router-controller: openapi_server.controllers.default_controller
components:
  schemas:
    Community:
      properties:
        id:
          type: integer
        about:
          type: string
        admins:
          items:
            type: string
          type: array
        avatar_url:
          type: string
        context:
          $ref: '#/components/schemas/CommunityContext'
        created_at:
          format: date-time
          type: string
        flag_text:
          type: string
        is_nsfw:
          type: boolean
        lang:
          type: string
        name:
          type: string
        num_authors:
          type: integer
        num_pending:
          type: integer
        settings:
          type: object
        subscribers:
          type: integer
        sum_pending:
          type: integer
        team:
          description: 'array of arrays with 3 strings: name, role_name, title'
          items:
            items:
              type: string
            maxItems: 3
            minItems: 3
            type: array
          type: array
        title:
          type: string
        type_id:
          type: integer
      required:
      - about
      - avatar_url
      - context
      - created_at
      - flag_text
      - id
      - is_nsfw
      - lang
      - name
      - num_authors
      - num_pending
      - settings
      - subscribers
      - sum_pending
      - team
      - title
      - type_id
      type: object
...
    AccountNotifications:
      description: 'array of objects with 6 notification related fields: date, id,
        msg, score, type and url'
      items:
        properties:
          date:
            description: date of the notification
            format: date-time
            type: string
          id:
            description: notification id
            type: integer
          msg:
            description: message of the notification
            type: string
          score:
            description: notification score
            type: integer
          type:
            description: 'type of notification: follow, vote, reply, reply_comment,
              mention, reblog'
            type: string
          url:
            description: end of notification url
            format: url
            type: string
        required:
        - date
        - id
        - msg
        - score
        - type
        - url
        type: object
      type: array
    DoesUserFollowAnyLists:
      description: answer to a question does user follow any list
      type: boolean
    GetAccountPosts:
      description: array of objects with posts related fields
      items:
        properties:
          active_votes:
            $ref: '#/components/schemas/ActiveVotes'
          author:
            description: account name of the post's author
            type: string
          author_payout_value:
            description: HBD paid to the author of the post
            type: string
          author_reputation:
            description: author's reputation score
            type: number
          beneficiaries:
            $ref: '#/components/schemas/Beneficiares'
          blacklists:
            items:
              description: account names of blacklisted users
              type: string
            type: array
          body:
            description: post content
            type: string
          category:
            description: post category
            type: string
          children:
            description: number of children comments
            type: integer
          created:
            description: creation date
            format: date-time
            type: string
          curator_payout_value:
            description: amount of HBD paid to curators
            type: string
          depth:
            description: nesting level
            type: integer
          is_paidout:
            description: information whether the post has been paid
            type: boolean
          json_metadata:
            type: object
          max_accepted_payout:
            description: maximal possible payout
            type: string
          net_rshares:
            description: netto rshares, result of rshares allocations
            type: integer
          payout:
            description: amount of payout
            type: number
          payout_at:
            description: date of payout
            format: date-time
            type: string
          pending_payout_value:
            description: pending or paid amount
            type: string
          percent_hbd:
            description: percent of HBD, 1000 = 100%
            type: integer
          permlink:
            description: post's permlink
            type: string
          post_id:
            description: id of the post, created from the author and the permlink
            type: integer
          promoted:
            description: amount of HBD if post is promoted
            type: string
          reblogged_by:
            items:
              description: account names of reblogers
              type: string
            type: array
          replies:
            items:
              description: replies
              type: string
            type: array
          stats:
            properties:
              flag_weight:
                type: number
              gray:
                type: boolean
              hide:
                type: boolean
              total_votes:
                type: integer
            required:
            - flag_weight
            - gray
            - hide
            - total_votes
            type: object
          title:
            description: post title
            type: string
          updated:
            description: date of update
            format: date-time
            type: string
          url:
            description: end of the url to the post, contains category, author and
              permlink
            type: string
        required:
        - active_votes
        - author
        - author_payout_value
        - author_reputation
        - beneficiaries
        - blacklists
        - body
        - category
        - children
        - created
        - curator_payout_value
        - depth
        - is_paidout
        - json_metadata
        - max_accepted_payout
        - net_rshares
        - payout
        - payout_at
        - pending_payout_value
        - percent_hbd
        - permlink
        - post_id
        - promoted
        - reblogged_by
        - replies
        - stats
        - title
        - updated
        - url
        type: object
      type: array
...
    ErrorMessage:
      properties:
        jsonrpc:
          type: string
        error:
          $ref: '#/components/schemas/ErrorMessage_error'
        id:
          type: integer
      required:
      - error
      - id
      - jsonrpc
      type: object
...
    CommunityRequest:
      example:
        method: bridge.get_community
        id: 0
        jsonrpc: "2.0"
        params:
          observer: good-karma
          name: hive-189306
      properties:
        jsonrpc:
          default: "2.0"
          type: string
        method:
          enum:
          - bridge.get_community
          type: string
        params:
          $ref: '#/components/schemas/CommunityRequest_params'
        id:
          default: 1
          type: integer
      required:
      - id
      - jsonrpc
      - method
      - params
      type: object
...
    AccountNotificationsRequest:
      example:
        method: bridge.account_notifications
        id: 5
        jsonrpc: "2.0"
        params:
          last_id: 6
          limit: 15
          account: blocktrades
          min_score: 8
      properties:
        jsonrpc:
          default: "2.0"
          type: string
        method:
          enum:
          - bridge.account_notifications
          type: string
        params:
          $ref: '#/components/schemas/AccountNotificationsRequest_params'
        id:
          default: 1
          type: integer
      required:
      - id
      - jsonrpc
      - method
      - params
      type: object
    DoesUserFollowAnyListsRequest:
      example:
        method: bridge.does_user_follow_any_lists
        id: 0
        jsonrpc: "2.0"
        params:
          observer: blocktrades
      properties:
        jsonrpc:
          default: "2.0"
          type: string
        method:
          enum:
          - bridge.does_user_follow_any_lists
          type: string
        params:
          $ref: '#/components/schemas/DoesUserFollowAnyListsRequest_params'
        id:
          default: 1
          type: integer
      required:
      - id
      - jsonrpc
      - method
      - params
      type: object
    GetAccountPostsRequest:
      example:
        method: bridge.get_account_posts
        id: 6
        jsonrpc: "2.0"
        params:
          observer: blocktrades
          limit: 8
          start_author: start_author
          sort: blog
          account: blocktrades
          start_permlink: start_permlink
      properties:
        jsonrpc:
          default: "2.0"
          type: string
        method:
          enum:
          - bridge.get_account_posts
          type: string
        params:
          $ref: '#/components/schemas/GetAccountPostsRequest_params'
        id:
          default: 1
          type: integer
      required:
      - id
      - jsonrpc
      - method
      - params
      type: object
...